requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
google-generativeai==0.3.2
PyGithub==2.1.1
chromadb==0.5.0
python-docx==1.1.0
PyPDF2==3.0.1
requests==2.31.0
Werkzeug==3.0.1
numpy<2.0
